requirements/base.txt
# Core Django and REST Framework
django==4.2.10  # Latest LTS release, compatible with most recent packages
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Django Extensions and Tools
django-celery-beat==2.7.0
django-debug-toolbar==4.4.6
django-extensions  # Remove version constraint for flexibility
django-redis  # Remove version constraint for flexibility
django-allauth[mfa]==65.0.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-cors-headers==4.4.0

# Misc Django Packages
django-environ==0.11.2
django-phonenumber-field>=8.0.0
django-model-utils==5.0.0
# Loosen version constraints for older packages to avoid conflicts
django-mptt
django-solo
django-fsm==2.7.1
django-imagekit==4.1.0
django-sequences
django-json-widget==1.1.1
django-admin-sortable2  # Remove version constraint for compatibility with Django 4.2
django-tinymce==3.4.0
django-log-request-id==2.0.0

# Utilities and Dependencies
python-slugify==8.0.4
Pillow==10.4.0
argon2-cffi==23.1.0
redis==5.1.1
hiredis==3.0.0
celery==5.4.0
psycopg2>=2.9.9
pytest-django==4.9.0
uvicorn==0.31.1
phonenumbers==8.12.21
pyotp==2.6.0
lxml==4.6.3
orjson==3.8.10
zeep==4.0.0
drf-writable-nested==0.6.2
xmltodict~=0.12.0
openpyxl==3.0.7
pdfkit==0.6.1
num2words~=0.5.10
drf-yasg~=1.21.7
drf-extra-fields~=3.4.0
pdf2image~=1.16.3
XlsxWriter==3.0.3
django-filter~=2.4.0
django-constance==4.1.3
PyPDF2==3.0.1
